Moving to our more common pattern of having a Models directory.
